requirements.txt
asttokens==2.2.1
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
decorator==4.4.2
executing==1.2.0
filelock==3.12.0
fonttools==4.39.3
idna==3.4
imageio==2.28.0
imageio-ffmpeg==0.4.8
importlib-resources==5.12.0
ipython==8.12.0
jedi==0.18.2
Jinja2==3.1.2
kiwisolver==1.4.4
lazy_loader==0.2
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
moviepy==1.0.3
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
proglog==0.1.10
prompt-toolkit==3.0.38
pure-eval==0.2.2
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.4.1
requests==2.29.0
scikit-image==0.20.0
scipy==1.9.1
six==1.16.0
stack-data==0.6.2
sympy==1.11.1
tifffile==2023.4.12
torch==2.0.0
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
zipp==3.15.0
